etl/requirements.txt
elasticsearch==8.5.3
python-dotenv==0.21.0
pydantic==1.10.4
psycopg2-binary==2.9.5